requirements.txt
anyio==3.6.2
blinker==1.6.2
Brotli==1.0.9
cachelib==0.9.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
dnspython==2.3.0
docopt==0.6.2
Flask==2.3.2
Flask-Caching==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mutagen==1.46.0
pipreqs==0.4.13
pycryptodomex==3.17
pymongo==4.3.3
pyngrok==6.0.0
pypng==0.20220715.0
PyQRCode==1.2.1
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
websockets==11.0.2
Werkzeug==2.3.3
yarg==0.1.9
yt-dlp==2023.3.4
zope.interface==6.0
